1) add create modal 2) add option on remote controller

+++ server/src/models/ScheduleEvent.js
@@ -3,7 +3,7 @@ const { DataTypes } = require('sequelize');
 const sequelize = require('../database');
 
 const ScheduleEvent = sequelize.define('ScheduleEvent', {
-  id: { type: DataTypes.INTEGER, primaryKey: true, autoIncrement: true },
+  id: { type: DataTypes.BIGINT, primaryKey: true, autoIncrement: true },
   scenario_id: { type: DataTypes.INTEGER, allowNull: false },
   event_order: { type: DataTypes.INTEGER, allowNull: false },
   start_time: { type: DataTypes.TIME, allowNull: false },
